Loss function uses output audio.
WaveNet uses input as output, but we're modifying it to use a different output.
Need to check on size of input and output to make sure they are the same or are offset and have padding as needed. Currently, an error is encountered after some training, because the logits size and labels size are not the same.